Add AnnotateRoutes::RemovalProcessor
